refactor: `components/common` 폴더 제거 #28
