+++ src/app/(mood)/waiting/funnels/step1/components/BodyTypeBottomSheet.tsx
@@ -1,13 +1,13 @@
 'use client';
 
-import Button from '@/components/common/Button/Button';
-import CheckBox from '@/components/common/CheckBox';
+import Button from '@/components/Button/Button';
+import CheckBox from '@/components/CheckBox';
 import { type UseFormReturn, useWatch } from 'react-hook-form';
 import type { MoodContextValue } from '../../../components/MoodContext';
-import BottomSheet from '@/components/common/Modal/BottomSheet';
+import BottomSheet from '@/components/Modal/BottomSheet';
 import useBottomSheet from '@/hooks/useBottomSheet';
-import { ButtonWrapper } from '@/components/common/Button';
-import Spacing from '@/components/common/Spacing';
+import { ButtonWrapper } from '@/components/Button';
+import Spacing from '@/components/Spacing';
 
 const BODY_TYPE_LIST = ['마름', '탄탄 슬림', '보통', '통통', '근육'] as const;
