+++ src/components/Header/TextHeader.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/components/common/Header';
+import { Header } from '@/components/Header';
 
 type TextHeaderProps = {
   text: string;
