+++ src/components/Modal/BottomDim.tsx
