+++ src/components/Modal/index.ts
@@ -1,0 +1,5 @@
+export { default as BottomDim } from './BottomDim';
+export { default as BottomSheet } from './BottomSheet';
+export { default as AnimatePortal } from './AnimatePortal';
+export { default as Portal } from './Portal';
+export { default as PopupContainer } from './PopupContainer';
