--- src/components/common/Modal/index.ts
@@ -1,2 +1,0 @@
-export { default as BottomDim } from './BottomDim';
-export { default as BottomSheet } from './BottomSheet';
