--- src/components/common/Portal/index.ts
@@ -1,2 +1,0 @@
-export { default as AnimatePortal } from './AnimatePortal';
-export { default as Portal } from './Portal';
